vmm: split vstate.rs in multiple files
Fixes #1605.
