+++ src/vmm/src/resources.rs
@@ -16,7 +16,7 @@ use crate::vmm_config::metrics::{init_metrics, MetricsConfig, MetricsConfigError
 use crate::vmm_config::mmds::{MmdsConfig, MmdsConfigError};
 use crate::vmm_config::net::*;
 use crate::vmm_config::vsock::*;
-use crate::vstate::VcpuConfig;
+use crate::vstate::vcpu::VcpuConfig;
 use mmds::ns::MmdsNetworkStack;
 use utils::net::ipv4addr::is_link_local_valid;
 
@@ -307,7 +307,7 @@ mod tests {
     use crate::vmm_config::net::{NetBuilder, NetworkInterfaceConfig};
     use crate::vmm_config::vsock::tests::default_config;
     use crate::vmm_config::RateLimiterConfig;
-    use crate::vstate::VcpuConfig;
+    use crate::vstate::vcpu::VcpuConfig;
     use dumbo::MacAddr;
     use logger::{LevelFilter, LOGGER};
     use utils::tempfile::TempFile;
