+++ src/vmm/src/vstate/mod.rs
@@ -1,0 +1,6 @@
+// Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+pub(crate) mod system;
+pub(crate) mod vcpu;
+pub(crate) mod vm;
