+++ src/vmm/src/vstate/system.rs
@@ -1,0 +1,114 @@
+// Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+//
+// Portions Copyright 2017 The Chromium OS Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the THIRD-PARTY file.
+
+use std::{
+    fmt::{Display, Formatter},
+    result,
+};
+
+use kvm_bindings::KVM_API_VERSION;
+use kvm_ioctls::Kvm;
+
+/// Errors associated with the wrappers over KVM ioctls.
+#[derive(Debug)]
+pub enum Error {
+    /// The host kernel reports an invalid KVM API version.
+    KvmApiVersion(i32),
+    /// Cannot initialize the KVM context due to missing capabilities.
+    KvmCap(kvm_ioctls::Cap),
+}
+
+impl Display for Error {
+    fn fmt(&self, f: &mut Formatter) -> std::fmt::Result {
+        use self::Error::*;
+
+        match self {
+            KvmApiVersion(v) => write!(
+                f,
+                "The host kernel reports an invalid KVM API version: {}",
+                v
+            ),
+            KvmCap(cap) => write!(f, "Missing KVM capabilities: {:?}", cap),
+        }
+    }
+}
+
+type Result<T> = result::Result<T, Error>;
+
+/// Describes a KVM context that gets attached to the microVM.
+/// It gives access to the functionality of the KVM wrapper as
+/// long as every required KVM capability is present on the host.
+pub struct KvmContext {
+    kvm: Kvm,
+    max_memslots: usize,
+}
+
+impl KvmContext {
+    pub fn new() -> Result<Self> {
+        use kvm_ioctls::Cap::*;
+        let kvm = Kvm::new().expect("Error creating the Kvm object");
+
+        // Check that KVM has the correct version.
+        if kvm.get_api_version() != KVM_API_VERSION as i32 {
+            return Err(Error::KvmApiVersion(kvm.get_api_version()));
+        }
+
+        // A list of KVM capabilities we want to check.
+        #[cfg(target_arch = "x86_64")]
+        let capabilities = vec![Irqchip, Ioeventfd, Irqfd, UserMemory, SetTssAddr];
+
+        #[cfg(target_arch = "aarch64")]
+        let capabilities = vec![Irqchip, Ioeventfd, Irqfd, UserMemory, ArmPsci02];
+
+        // Check that all desired capabilities are supported.
+        match capabilities
+            .iter()
+            .find(|&capability| !kvm.check_extension(*capability))
+        {
+            None => {
+                let max_memslots = kvm.get_nr_memslots();
+                Ok(KvmContext { kvm, max_memslots })
+            }
+
+            Some(c) => Err(Error::KvmCap(*c)),
+        }
+    }
+
+    pub fn fd(&self) -> &Kvm {
+        &self.kvm
+    }
+
+    /// Get the maximum number of memory slots reported by this KVM context.
+    pub fn max_memslots(&self) -> usize {
+        self.max_memslots
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::fs::File;
+
+    use super::*;
+
+    #[test]
+    fn test_kvm_context() {
+        use std::os::unix::fs::MetadataExt;
+        use std::os::unix::io::{AsRawFd, FromRawFd};
+
+        let c = KvmContext::new().unwrap();
+
+        assert!(c.max_memslots() >= 32);
+
+        let kvm = Kvm::new().unwrap();
+        let f = unsafe { File::from_raw_fd(kvm.as_raw_fd()) };
+        let m1 = f.metadata().unwrap();
+        let m2 = File::open("/dev/kvm").unwrap().metadata().unwrap();
+
+        assert_eq!(m1.dev(), m2.dev());
+        assert_eq!(m1.ino(), m2.ino());
+    }
+}
